Add font settings to config dialog
The fonts provided in the combo box are all monospaced fonts. Though the
user can enter whatever name of a font he wants. If it's not monospace
though, he might be unhappy with the result. If the font name typed in
isn't known, it seems her defaults to times new roman or something.
Which obviously isn't monospace.
The combo box is filled dynamically. It
searches all installed fonts and checks manually if they are monospaced.
Which takes a while when first opening the combo box.